add report

+++ SSU.DepManagement/Models/View/AppFormDisciplineViewItem.cs
@@ -2,13 +2,15 @@
 
 public class AppFormDisciplineViewItem
 {
+    
+    
     public string Name { get; set; }
     
     public int Semester { get; set; }
 
     public IReadOnlyList<int> Groups { get; set; }
     
-    public int TotalHours { get; set; }
+    public double TotalHours { get; set; }
     
     public IReadOnlyList<RequestViewItem> Requests { get; set; }
 }
